chore: move monacoEnv to app/core
